# openapi.json
{
  "components": {
    "schemas": {
...
      "ChoiceResult": {
        "additionalProperties": false,
        "properties": {
          "abstain_probability": {
            "title": "Abstain Probability",
            "type": "number"
          },
          "actionable": {
            "description": "Whether the answer cleared the operating floor.",
            "title": "Actionable",
            "type": "boolean"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The option id, or None when no option applies.",
            "title": "Answer"
          },
          "certificate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Certificate"
              },
              {
                "type": "null"
              }
            ]
          },
          "certified": {
            "description": "Whether the answer passed the policy's threshold at the requested risk.",
            "title": "Certified",
            "type": "boolean"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Calibrated probability of the answer, compared with the floor.",
            "title": "Confidence"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "prediction_set": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionSet"
              },
              {
                "type": "null"
              }
            ]
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Probabilities",
            "type": "object"
          },
          "relevant_context": {
            "items": {
              "$ref": "#/components/schemas/ContextRelevance"
            },
            "title": "Relevant Context",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "type": {
            "const": "choice",
            "default": "choice",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "status",
          "actionable",
          "certified",
          "confidence",
          "relevant_context",
          "certificate",
          "latency_ms",
          "answer",
          "probabilities",
          "abstain_probability",
          "prediction_set"
        ],
        "title": "ChoiceResult",
        "type": "object"
      },
...
      "DecideRequest": {
        "additionalProperties": false,
        "description": "Body of `POST /v1/decide`. `alpha=None` uses the release default.",
        "properties": {
          "alpha": {
            "anyOf": [
              {
                "exclusiveMaximum": 1.0,
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alpha"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "$ref": "#/components/schemas/Context"
              },
              {
                "$ref": "#/components/schemas/JsonState"
              }
            ],
            "description": "A string (one passage), a list of strings (passages), a Context object, or {'state': <JSON>} for a structured state.",
            "title": "Context"
          },
          "decision": {
            "discriminator": {
              "mapping": {
                "choice": "#/components/schemas/Choice",
                "estimate": "#/components/schemas/Estimate",
                "multi_choice": "#/components/schemas/MultiChoice",
                "rank": "#/components/schemas/Rank",
                "rate": "#/components/schemas/Rate",
                "verify": "#/components/schemas/Verify",
                "yes_no": "#/components/schemas/YesNo"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Choice"
              },
              {
                "$ref": "#/components/schemas/MultiChoice"
              },
              {
                "$ref": "#/components/schemas/YesNo"
              },
              {
                "$ref": "#/components/schemas/Verify"
              },
              {
                "$ref": "#/components/schemas/Rank"
              },
              {
                "$ref": "#/components/schemas/Rate"
              },
              {
                "$ref": "#/components/schemas/Estimate"
              }
            ],
            "title": "Decision"
          },
          "min_confidence": {
            "anyOf": [
              {
                "exclusiveMaximum": 1.0,
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Confidence"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode",
            "default": "standard"
          },
          "risk": {
            "default": 0.01,
            "title": "Risk",
            "type": "number"
          }
        },
        "required": [
          "context",
          "decision"
        ],
        "title": "DecideRequest",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "description": "Body of every error response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "Mode": {
        "enum": [
          "standard",
          "threshold",
          "certified"
        ],
        "title": "Mode",
        "type": "string"
      },
...
      "MultiChoiceResult": {
        "additionalProperties": false,
        "properties": {
          "actionable": {
            "description": "Whether the answer cleared the operating floor.",
            "title": "Actionable",
            "type": "boolean"
          },
          "answer": {
            "items": {
              "type": "string"
            },
            "title": "Answer",
            "type": "array"
          },
          "certificate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Certificate"
              },
              {
                "type": "null"
              }
            ]
          },
          "certified": {
            "description": "Whether the answer passed the policy's threshold at the requested risk.",
            "title": "Certified",
            "type": "boolean"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Calibrated probability of the answer, compared with the floor.",
            "title": "Confidence"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Independent probability that each option applies.",
            "title": "Probabilities",
            "type": "object"
          },
          "relevant_context": {
            "items": {
              "$ref": "#/components/schemas/ContextRelevance"
            },
            "title": "Relevant Context",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "type": {
            "const": "multi_choice",
            "default": "multi_choice",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "status",
          "actionable",
          "certified",
          "confidence",
          "relevant_context",
          "certificate",
          "latency_ms",
          "answer",
          "probabilities"
        ],
        "title": "MultiChoiceResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Typed, calibrated and certified decisions.",
    "title": "MIMIR",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/decide": {
      "post": {
        "operationId": "decide_v1_decide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "choice": "#/components/schemas/ChoiceResult",
                      "estimate": "#/components/schemas/EstimateResult",
                      "multi_choice": "#/components/schemas/MultiChoiceResult",
                      "rank": "#/components/schemas/RankResult",
                      "rate": "#/components/schemas/RateResult",
                      "verify": "#/components/schemas/VerifyResult",
                      "yes_no": "#/components/schemas/YesNoResult"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChoiceResult"
                    },
                    {
                      "$ref": "#/components/schemas/MultiChoiceResult"
                    },
                    {
                      "$ref": "#/components/schemas/YesNoResult"
                    },
                    {
                      "$ref": "#/components/schemas/VerifyResult"
                    },
                    {
                      "$ref": "#/components/schemas/RankResult"
                    },
                    {
                      "$ref": "#/components/schemas/RateResult"
                    },
                    {
                      "$ref": "#/components/schemas/EstimateResult"
                    }
                  ],
                  "title": "Response Decide V1 Decide Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Decide"
      }
    },
    "/v1/decide/batch": {
      "post": {
        "operationId": "decide_batch_v1_decide_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Decide Batch"
      }
    },
    "/v1/decide/uncertified": {
      "post": {
        "operationId": "decide_uncertified_v1_decide_uncertified_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UncertifiedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "choice": "#/components/schemas/ChoiceResult",
                      "estimate": "#/components/schemas/EstimateResult",
                      "multi_choice": "#/components/schemas/MultiChoiceResult",
                      "rank": "#/components/schemas/RankResult",
                      "rate": "#/components/schemas/RateResult",
                      "verify": "#/components/schemas/VerifyResult",
                      "yes_no": "#/components/schemas/YesNoResult"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChoiceResult"
                    },
                    {
                      "$ref": "#/components/schemas/MultiChoiceResult"
                    },
                    {
                      "$ref": "#/components/schemas/YesNoResult"
                    },
                    {
                      "$ref": "#/components/schemas/VerifyResult"
                    },
                    {
                      "$ref": "#/components/schemas/RankResult"
                    },
                    {
                      "$ref": "#/components/schemas/RateResult"
                    },
                    {
                      "$ref": "#/components/schemas/EstimateResult"
                    }
                  ],
                  "title": "Response Decide Uncertified V1 Decide Uncertified Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Decide Uncertified"
      }
    },
    "/v1/models": {
      "get": {
        "operationId": "models_v1_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Models"
      }
    },
    "/v1/systemone": {
      "post": {
        "operationId": "systemone_v1_systemone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemOneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Systemone"
      }
    },
    "/v1/tools/{name}": {
      "post": {
        "operationId": "call_tool_v1_tools__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolArguments"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "choice": "#/components/schemas/ChoiceResult",
                      "estimate": "#/components/schemas/EstimateResult",
                      "multi_choice": "#/components/schemas/MultiChoiceResult",
                      "rank": "#/components/schemas/RankResult",
                      "rate": "#/components/schemas/RateResult",
                      "verify": "#/components/schemas/VerifyResult",
                      "yes_no": "#/components/schemas/YesNoResult"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChoiceResult"
                    },
                    {
                      "$ref": "#/components/schemas/MultiChoiceResult"
                    },
                    {
                      "$ref": "#/components/schemas/YesNoResult"
                    },
                    {
                      "$ref": "#/components/schemas/VerifyResult"
                    },
                    {
                      "$ref": "#/components/schemas/RankResult"
                    },
                    {
                      "$ref": "#/components/schemas/RateResult"
                    },
                    {
                      "$ref": "#/components/schemas/EstimateResult"
                    }
                  ],
                  "title": "Response Call Tool V1 Tools  Name  Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Call Tool"
      }
    }
  }
}